the beginning of code decomposition and its adaptation to rtos
