+++ include/periphery.h
@@ -1,0 +1,9 @@
+#ifndef PERIPH_INIT_H
+#define PERIPH_INIT_H
+
+void clock_setup(void);
+void gpio_setup(void);
+void adc_setup(void);
+
+#endif // PERIPH_INIT_H
+
