+++ src/periphery.c
@@ -1,0 +1,50 @@
+#include <libopencm3/stm32/f4/rcc.h>
+#include <libopencm3/stm32/f4/gpio.h>
+#include <libopencm3/stm32/f4/adc.h>
+
+void clock_setup(void) {
+    rcc_osc_on(RCC_HSE);
+    rcc_wait_for_osc_ready(RCC_HSE);
+    rcc_set_sysclk_source(RCC_HSE);
+
+    rcc_periph_clock_enable(RCC_GPIOB);
+    rcc_periph_clock_enable(RCC_I2C1);
+    rcc_periph_clock_enable(RCC_GPIOA);
+    rcc_periph_clock_enable(RCC_USART2);
+}
+
+void adc_setup(void) {
+    rcc_periph_clock_enable(RCC_ADC1);
+    rcc_periph_clock_enable(RCC_GPIOA);
+
+    gpio_mode_setup(GPIOA, GPIO_MODE_ANALOG, GPIO_PUPD_NONE, GPIO0);
+
+    adc_power_off(ADC1);
+    adc_disable_scan_mode(ADC1);
+    adc_set_single_conversion_mode(ADC1);:
+   
+    uint8_t channels[1] = { ADC_CHANNEL0 };
+    adc_set_regular_sequence(ADC1, 1, channels);
+
+    adc_set_sample_time(ADC1, ADC_CHANNEL0, ADC_SMPR_SMP_28CYC);
+    adc_power_on(ADC1);
+    dwt_delay_us(10);
+}
+
+void gpio_setup(void) {
+    gpio_mode_setup(GPIOB, GPIO_MODE_INPUT, GPIO_PUPD_NONE, GPIO10);
+    gpio_mode_setup(GPIOB, GPIO_MODE_AF, GPIO_PUPD_PULLUP, GPIO8 | GPIO9);
+    gpio_set_output_options(GPIOB, GPIO_OTYPE_OD, GPIO_OSPEED_50MHZ, GPIO8 | GPIO9);
+    gpio_set_af(GPIOB, GPIO_AF4, GPIO8 | GPIO9);
+
+    gpio_mode_setup(GPIOA, GPIO_MODE_AF, GPIO_PUPD_NONE, GPIO2 | GPIO3);
+    gpio_set_af(GPIOA, GPIO_AF7, GPIO2 | GPIO3);
+
+    gpio_mode_setup(GPIOD, GPIO_MODE_OUTPUT, GPIO_PUPD_NONE, GREEN_LED);
+    gpio_clear(GPIOD, GREEN_LED);
+
+}   
+
+
+
+
